requirements.txt
aiohttp==3.6.2
alpaca-trade-api==0.48
alpha-vantage==2.2.0
appdirs==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
asyncio-nats-client==0.10.0
atomicwrites==1.4.0
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
autopep8==1.5.4
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.4
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==4.0.2
constantly==15.1.0
coverage==5.0.4
crayons==0.3.0
cryptography==3.2
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
discord==1.0.1
discord.py==1.5.1
distlib==0.3.1
Django==3.1.2
docutils==0.16
entrypoints==0.3
enum34==1.1.10
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
google==3.0.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.1
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-tabnine==1.1.0
keyring==21.2.1
kiteconnect==3.8.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.4.0
MouseInfo==0.1.2
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.1.5
numpy==1.18.5
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.1.0
pipenv==2020.8.13
pkginfo==1.5.0.1
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2==2.8.6
py==1.9.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyGetWindow==0.0.8
Pygments==2.6.1
PyHamcrest==2.0.2
PyMsgBox==1.0.7
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.7.0
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
PyTweening==1.0.3
pytz==2020.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
readme-renderer==26.0
requests==2.23.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
selenium==3.141.0
selenium-testing-module==1.0.3
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
twine==3.2.0
Twisted==20.3.0
txaio==20.4.1
ujson==3.2.0
uritemplate==3.0.1
urllib3==1.24.3
virtualenv==20.0.31
virtualenv-clone==0.5.4
wcwidth==0.1.9
webdriver-manager==2.3.0
webdrivermanager==0.7.4
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.4.2
zope.interface==5.1.0
